requirements.txt
python-dotenv
requests
pillow-heif
Pillow
ffmpeg-python
PyYAML>=6.0.1  # For YAML configuration file support